HW3

+++ hw3/matrixNormCuda3.cu
@@ -222,9 +222,9 @@ void matrixNorm() {
 
   float **h_sums;
   //Get user input into size;
-  h_sums = new int*[BLOCK_SIZE];
+  h_sums = new float*[BLOCK_SIZE];
   for (int i = 0; i < BLOCK_SIZE; i++)
-     h_sums[i] = new int[BLOCK_SIZE];
+     h_sums[i] = new float[BLOCK_SIZE];
 
   for (int i=0; i < BLOCK_SIZE; i++)
       for (int j=0; j < BLOCK_SIZE; j++)
